重力ページめくり ボタンページめくり スワイプページめくり を追加・設定できるようにした。 はじめからを選択したときに読めないファイルを読んでいたのを無くした。 ヘルプを追加した。

+++ trunk/Application.h
@@ -45,6 +45,6 @@
 	ImageView* _imageview;
 }
 
--(void) gravity: (int)x  gy:(int) y gz:(int)z;
+-(void) gravity: (float)x  gy:(float) y gz:(float)z;
 @end
